refactor: remove compatibility re-exports, all paths canonical
