+++ src/main.rs
@@ -5,8 +5,9 @@ use color_eyre::eyre::Result;
 use tokio::sync::oneshot;
 use tracing::info;
 use twitch_1337::{
-    Services, aviation, ensure_data_dir, get_data_dir, install_crypto_provider, install_tracing,
-    llm, load_configuration, run_bot, setup_and_verify_twitch_client, util::clock::SystemClock,
+    Services, ai::llm, aviation, ensure_data_dir, get_data_dir, install_crypto_provider,
+    install_tracing, load_configuration, run_bot, setup_and_verify_twitch_client,
+    util::clock::SystemClock,
 };
 
 #[tokio::main]
